requirements.txt
Flask
SQLAlchemy
pyodbc
python-dotenv
pandas
requests
openpyxl
